Question Juggling Framework
Updated the questions to look more like juggling. Added a startup and reworked how they spawn to allow for different number of responses. Added a timer to questions that forces a bad response if nothing is chosen in time. Updated TempoMananger to remain between scenes, and updated the tempo variable to be static.